# openapi.yaml
openapi: 3.0.3
info:
  title: Rated — Music Rating & Listening Stats API
  version: 0.1.0
  description: |
    API for a Letterboxd-style app for music: users rate/log/review tracks,
    albums and artists, and link streaming accounts (Spotify, Apple Music,
    Amazon Music, YouTube Music) to pull personal listening stats (top
    artists/tracks over time, estimated listening time) into their profile.

    ## Design principles

    1. **Ratings are platform-agnostic.** A rating is attached to a canonical
       `Artist` / `Album` / `Track` entity owned by this app, never directly to
       a Spotify/Apple/Amazon/YouTube ID. This is required because the same
       user may rate an album they streamed on Spotify while a friend rates
       the same album via Apple Music — they must resolve to the same row.
       Canonical entities are seeded/matched via ISRC/UPC and a metadata
       source such as MusicBrainz, then enriched with `external_ids` per
       provider (`{ "spotify": "...", "apple_music": "..." }`) so the client
       can deep-link "open in Spotify" / "open in Apple Music" buttons.

    2. **Streaming accounts are OAuth connections, not logins.** A user signs
       up with the app directly (email/password or its own OAuth), then
       *optionally* links one or more streaming accounts under `/connections`
       purely to import listening data. Losing a streaming connection must
       never lose ratings or reviews.

...

...

       Consequences baked into this spec:
...

...
servers:
...
paths:
...
  /diary:
    post:
      tags: [Diary]
      summary: Log a listen (optionally with a rating), Letterboxd-diary style
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DiaryEntryCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiaryEntry' }
  /users/{username}/diary:
    get:
      tags: [Diary]
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/DiaryEntry' } }
                  page: { type: integer }
                  page_size: { type: integer }
                  total: { type: integer }
  /diary/{id}:
    delete:
      tags: [Diary]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '204': { description: Deleted }

  # -------------------------------------------------------------- Stats
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    UsernamePath:
      name: username
      in: path
      required: true
      schema: { type: string }
...
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
...
  schemas:
...
    PublicUser:
      type: object
      properties:
        username: { type: string }
        display_name: { type: string }
        avatar_url: { type: string, format: uri, nullable: true }
        bio: { type: string, nullable: true }
...
    AlbumSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        cover_url: { type: string, format: uri, nullable: true }
    ArtistSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
...
    SubjectType:
      type: string
      enum: [track, album, artist]
    Score:
      type: number
      minimum: 0.5
      maximum: 5
      multipleOf: 0.5
      description: Half-star scale, 0.5–5.0 (Letterboxd-style)
...
    Rating:
      type: object
      properties:
        id: { type: string, format: uuid }
        user: { $ref: '#/components/schemas/PublicUser' }
        subject_type: { $ref: '#/components/schemas/SubjectType' }
        subject:
          description: Polymorphic summary of the rated artist/album/track
          oneOf:
            - $ref: '#/components/schemas/ArtistSummary'
            - $ref: '#/components/schemas/AlbumSummary'
        score: { $ref: '#/components/schemas/Score' }
        review: { type: string, nullable: true }
        is_relisten: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    DiaryEntryCreateRequest:
      type: object
      required: [subject_type, subject_id, listened_on]
      properties:
        subject_type: { $ref: '#/components/schemas/SubjectType' }
        subject_id: { type: string, format: uuid }
        listened_on: { type: string, format: date }
        score: { $ref: '#/components/schemas/Score' }
        review: { type: string, maxLength: 5000, nullable: true }
    DiaryEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        subject_type: { $ref: '#/components/schemas/SubjectType' }
        subject:
          oneOf:
            - $ref: '#/components/schemas/ArtistSummary'
            - $ref: '#/components/schemas/AlbumSummary'
        rating: { $ref: '#/components/schemas/Rating', nullable: true }
        listen_count:
          type: integer
          description: >-
            How many diary entries this user has logged for this exact
            subject, this one included - i.e. how many times they've logged
            listening to it. Derived entirely from the user's own diary, not
            from a streaming provider (no provider API exposes per-track
            lifetime play counts).
        listened_on: { type: string, format: date }
        created_at: { type: string, format: date-time }
...